move user mgmt func to pro
