+++ pro/initialize.go
@@ -22,6 +22,7 @@ import (
 func InitPro() {
 	servercfg.IsPro = true
 	models.SetLogo(retrieveProLogo())
+	proLogic.UserRolesInit()
 	controller.HttpMiddlewares = append(
 		controller.HttpMiddlewares,
 		proControllers.OnlyServerAPIWhenUnlicensedMiddleware,
@@ -124,6 +125,10 @@ func InitPro() {
 	logic.NetworkPermissionsCheck = proLogic.NetworkPermissionsCheck
 	logic.GlobalPermissionsCheck = proLogic.GlobalPermissionsCheck
 	logic.DeleteNetworkRoles = proLogic.DeleteNetworkRoles
+	logic.CreateDefaultNetworkRoles = proLogic.CreateDefaultNetworkRoles
+	logic.FilterNetworksByRole = proLogic.FilterNetworksByRole
+	logic.IsGroupsValid = proLogic.IsGroupsValid
+	logic.RemoveNetworkRoleFromUsers = proLogic.RemoveNetworkRoleFromUsers
 }
 
 func retrieveProLogo() string {
